Sagar_Dataharvesting/Real-Time DB/data_harvesting 2.0/requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
python-socketio==5.11.4
python_dateutil==2.8.2
redis==5.0.3
Requests==2.32.3
